AI_Resume_Builder/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
together
google-generativeai
python-dotenv
jinja2
